backend/requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-import-export==3.3.7
django-jazzmin==3.0.0
django-sortedm2m==3.1.1
djangorestframework==3.14.0
et_xmlfile==2.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.5
pillow==11.0.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
